requirements.txt
aniso8601==1.2.0
argcomplete==1.8.2
asn1crypto==0.22.0
base58==0.2.4
beautifulsoup4==4.6.0
boto3==1.4.4
botocore==1.5.40
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
click==6.7
cryptography==1.9
docutils==0.13.1
durationpy==0.4
enum34==1.1.6
Flask==0.12.1
Flask-Ask==0.9.3
future==0.16.0
futures==3.1.1
hjson==2.0.2
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.2
kappa==0.6.0
lambda-packages==0.15.0
MarkupSafe==1.0
placebo==0.8.1
pycparser==2.17
pyOpenSSL==17.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-slugify==1.2.4
PyYAML==3.12
requests==2.17.3
s3transfer==0.1.10
six==1.10.0
toml==0.9.2
tqdm==4.11.2
troposphere==1.9.3
Unidecode==0.4.20
urllib3==1.21.1
Werkzeug==0.12
wikipedia==1.4.0
wsgi-request-logger==0.4.6
zappa==0.42.0
